pyproject.toml
[project]
name = "task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.9",
    "google-genai>=1.29.0",
    "lancedb>=0.24.2",
    "pypdf>=5.9.0",
    "python-docx>=1.2.0",
    "rapidocr-onnxruntime>=1.4.4",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
]
